# gin/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/auth/google/token": {
            "get": {
                "description": "2000 성공\n2001 신규 유저 닉네임 입력 화면으로 이동\n4004 required token\n4003 not verify id token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "구글 idtoken",
                "parameters": [
                    {
                        "type": "string",
                        "description": "idtoken",
                        "name": "token",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/auth.UserResponse"
                        }
                    }
                }
            }
        },
...
        "/auth/token": {
            "post": {
                "description": "2000 성공\n4001 required refresh_token\n4101 required login\n4004 다른 웹에서 로그인되어 있습니다 다시 로그인 해주세요",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "토큰",
                "parameters": [
                    {
                        "description": "refresh token",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.TokenBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/auth.TokenResponse"
                        }
                    }
                }
            }
        },
...
        "/interest": {
            "get": {
                "description": "2000 성공",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "interest"
                ],
                "summary": "관심종목",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "offset count",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/interest.InterestListResponse"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "2000 성공\n4001 missing bodies\n4002 cannot add ticker\n4101 required login",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "interest"
                ],
                "summary": "관심종목 추가",
                "parameters": [
                    {
                        "description": "add interest",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/interest.SetInterestBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/interest.InterestListResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "2000 성공\n4001 missing bodies\n4002 cannot delete[sql]\n4101 required login",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "interest"
                ],
                "summary": "관심종목 삭제",
                "parameters": [
                    {
                        "description": "delete interest",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/interest.DeleteIntereestBody"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/interest/sale": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "2000 성공\n4002 missing bodies\n4101 required login\n4004 Cannot sale\n4005 The time to sale has over",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "interest"
                ],
                "summary": "관심종목 매도",
                "parameters": [
                    {
                        "description": "sale interest",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/interest.SaleInterestBody"
                        }
                    }
                ],
                "responses": {}
            }
        },
...
        "/ticker/comment": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "2000 성공\n4101 required login\n4003 empty comment or code\n4004 missing bodies",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "comment"
                ],
                "summary": "티커 코멘트 등록",
                "parameters": [
                    {
                        "description": "ticker comment",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/comment.NewCommentBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/comment.TickerCommentsResponse"
                            }
                        }
                    }
                }
            }
        },
        "/ticker/comment/{ticker}": {
            "get": {
                "description": "2000 성공\n4101 required login\n4004 count , created_at error",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "comment"
                ],
                "summary": "티커 코멘트 목록",
                "parameters": [
                    {
                        "type": "string",
                        "description": "code",
                        "name": "ticker",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "count",
                        "name": "count",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "created_at",
                        "name": "created_at",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/comment.TickerCommentsResponse"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "auth.TokenBody": {
            "type": "object",
            "properties": {
                "refresh_token": {
                    "type": "string"
                }
            }
        },
        "auth.TokenResponse": {
            "type": "object",
            "properties": {
                "access_token": {
                    "description": "access token",
                    "type": "string"
                },
                "refresh_token": {
                    "description": "refresh token",
                    "type": "string"
                }
            }
        },
        "auth.UserResponse": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "created_at": {
                    "description": "가입일",
                    "type": "string"
                },
                "email": {
                    "description": "email 마우스 호버시에만 보이게",
                    "type": "string"
                },
                "nickname": {
                    "description": "nickname",
                    "type": "string"
                },
                "profile_image": {
                    "description": "google login시 받아오는 profile image",
                    "type": "string"
                },
                "refresh_token": {
                    "type": "string"
                }
            }
        },
        "comment.NewCommentBody": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "comment": {
                    "type": "string"
                }
            }
        },
        "comment.TickerCommentsResponse": {
            "type": "object",
            "properties": {
                "comment": {
                    "description": "코멘트",
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/comment.commentUser"
                }
            }
        },
        "comment.commentUser": {
            "type": "object",
            "properties": {
                "nickname": {
                    "description": "닉네임",
                    "type": "string"
                },
                "profile_image": {
                    "description": "프로필이미지",
                    "type": "string"
                }
            }
        },
        "interest.DeleteIntereestBody": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                }
            }
        },
        "interest.InterestListResponse": {
            "type": "object",
            "properties": {
                "recent": {
                    "$ref": "#/definitions/interest.interestTickerChartModel"
                },
                "ticker": {
                    "$ref": "#/definitions/interest.interestList"
                }
            }
        },
        "interest.SaleInterestBody": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                }
            }
        },
        "interest.SetInterestBody": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                }
            }
        },
        "interest.interestList": {
            "type": "object",
            "properties": {
                "close": {
                    "description": "등록한 시점의 가격",
                    "type": "number"
                },
                "date_time": {
                    "description": "등록된 시점",
                    "type": "string"
                },
                "market": {
                    "description": "0 KOSPI 1 KOSDAQ",
                    "type": "integer"
                },
                "name": {
                    "description": "종목이름",
                    "type": "string"
                },
                "percent": {
                    "description": "등록한 시점의 등락률",
                    "type": "string"
                },
                "symbol": {
                    "description": "종목코드",
                    "type": "string"
                },
                "ticker": {
                    "$ref": "#/definitions/interest.interestTickerModel"
                },
                "type": {
                    "description": "현재 사용하지 않습니다.",
                    "type": "integer"
                },
                "user": {
                    "$ref": "#/definitions/interest.interestUserModel"
                },
                "volume": {
                    "description": "등록된 시점의 거래량",
                    "type": "string"
                }
            }
        },
        "interest.interestTickerChartModel": {
            "type": "object",
            "properties": {
                "close": {
                    "description": "현재가",
                    "type": "number"
                },
                "percent": {
                    "description": "현재 등락률",
                    "type": "string"
                },
                "volume": {
                    "description": "현재 거래량",
                    "type": "string"
                }
            }
        },
        "interest.interestTickerModel": {
            "type": "object",
            "properties": {
                "homepage": {
                    "description": "홈페이지",
                    "type": "string"
                }
            }
        },
        "interest.interestUserModel": {
            "type": "object",
            "properties": {
                "nickname": {
                    "description": "등록한 사용자의 닉네임",
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}